delete from_substrate_issuing::Config from precompile
