+++ node/runtime/pangolin/src/pallets/from_substrate_issuing_.rs
@@ -4,7 +4,7 @@ use bp_runtime::ChainId;
 use frame_support::PalletId;
 use sp_runtime::AccountId32;
 // --- darwinia-network ---
-use crate::{pangoro_messages::ToPangoroOutboundPayLoad, *};
+use crate::{pangoro_messages::ToPangoroMessageSender, *};
 use darwinia_support::{s2s::ToEthAddress, ChainName};
 use drml_bridge_primitives::{AccountIdConverter, PANGORO_CHAIN_ID, PANGORO_PANGOLIN_LANE};
 use from_substrate_issuing::Config;
@@ -34,7 +34,7 @@ impl Config for Runtime {
 	type BridgedAccountIdConverter = AccountIdConverter;
 	type BridgedChainId = PangoroChainId;
 	type ToEthAddressT = TruncateToEthAddress;
-	type OutboundPayloadCreator = ToPangoroOutboundPayLoad;
+	type OutboundPayloadCreator = ToPangoroMessageSender;
 	type InternalTransactHandler = Ethereum;
 	type BackingChainName = BackingChainName;
 	type MessageLaneId = BridgePangoroLaneId;
