+++ node/runtime/pangoro/src/pallets/to_substrate_backing_.rs
@@ -42,8 +42,7 @@ impl Config for Runtime {
 	type RingCurrency = Ring;
 	type BridgedAccountIdConverter = AccountIdConverter;
 	type BridgedChainId = PangolinChainId;
-	// TODO
-	//type InternalTransactHandler = Ethereum;
+	type InternalTransactHandler = Ethereum;
 	type OutboundPayloadCreator = ToPangolinOutboundPayload;
 	type MessageNoncer = PangolinMessageNoncer;
 	type MessageLaneId = BridgePangolinLaneId;
